feat: Reduce bundle size, extract parser & utils, and fix typings
