+++ src/const.ts
@@ -1,0 +1,2 @@
+export const APPLICATION_JSON = 'application/json';
+export const CONTENT_TYPE = 'Content-Type';
